refactor pric double to bigdecimal

+++ src/main/java/com/os/yerinial/model/dto/customer/response/GetCustomerReservationResponse.java
@@ -2,12 +2,14 @@ package com.os.yerinial.model.dto.customer.response;
 
 import com.os.yerinial.model.entity.ReservationStatus;
 
+import java.math.BigDecimal;
+
 public record GetCustomerReservationResponse(
         Long reservationId,
         Long eventId,
         Long customerId,
         String eventName,
-        double totalPrice,
+        BigDecimal totalPrice,
         int ticketCount,
         ReservationStatus status
 ) {
